+++ src/ambition/7kaaInterface/draw.cc
@@ -31,13 +31,16 @@
 #include "OBUTT3D.h"
 #include "OBUTTCUS.h"
 #include "OCONFIG.h"
+#include "OF_BASE.h"
 #include "OF_HARB.h"
+#include "OF_MONS.h"
 #include "OF_WAR.h"
 #include "OFIRM.h"
 #include "OFONT.h"
 #include "OIMGRES.h"
 #include "ONATIONA.h"
 #include "OPOWER.h"
+#include "ORACERES.h"
 #include "OTOWN.h"
 #include "OUNIT.h"
 #include "vga_util.h"
@@ -54,6 +57,44 @@
 
 
 namespace _7kaaAmbitionInterface::Draw {
+
+const std::map<
+  int,
+  Ambition::UserInterface::Point
+> BUILDING_DRAWING_OFFSETS = {
+  { 11, { 0, -8 } }, /* Fort. */
+  { 12, { 2, -10 } }, /* Factory. */
+  { 17, { 0, -12 } }, /* Inn. */
+  { 18, { 0, -12 } }, /* Market. */
+  { 19, { 0, -6 } }, /* Mine. */
+  { 33, { 2, -12 } }, /* Tower of Science. */
+  { 34, { 0, -12 } }, /* War Factory. */
+
+  { 1, { 1, -22 } }, /* Norman Seat of Power. */
+  { 2, { 0, -20 } }, /* Mayan Seat of Power. */
+  { 3, { 0, -10 } }, /* Greek Seat of Power. */
+  { 4, { 0, -20 } }, /* Viking Seat of Power. */
+  { 5, { 0, -22 } }, /* Persian Seat of Power. */
+  { 6, { 0, -20 } }, /* Chinese Seat of Power. */
+  { 7, { 0, -22 } }, /* Japanese Seat of Power. */
+  { 8, { 5, -20 } }, /* Egyptian Seat of Power. */
+  { 9, { 0, -12 } }, /* Mughal Seat of Power. */
+  { 10, { 0, -14 } }, /* Zulu Seat of Power. */
+
+  { 20, { -6, -14 } }, /* Holgh Lair. */
+  { 21, { 4, -10 } }, /* Karrotten Lair. */
+  { 22, { 0, -16 } }, /* Haubudam Lair. */
+  { 23, { -4, -2 } }, /* Deezboanz Lair. */
+  { 24, { -2, -10 } }, /* Droog Lair. */
+  { 25, { 0, -18 } }, /* Wyrm Lair. */
+  { 26, { -4, -8 } }, /* Doink Lair. */
+  { 27, { 1, -6 } }, /* Brooksen Lair. */
+  { 28, { 0, -8 } }, /* Pfith Lair. */
+  { 29, { 0, -8 } }, /* Rattus Lair. */
+  { 30, { -10, -30 } }, /* Ick Lair. */
+  { 31, { -2, -8 } }, /* Sauroid Lair. */
+  { 32, { -1, -8 } }, /* Rokken Lair. */
+};
 
 const std::map<short, short> ROCK_BITMAP_REMAPPINGS = {
   /* == Look like they should block but do not. == */
@@ -723,6 +764,26 @@ FirmBitmap* calculateFirmBitmap(
   }
 
   return Ambition::calculateFirmBitmap(_7kaaCalculation, firm);
+}
+
+void calculateFirmDrawLocation(
+  const int _7kaaFirmBuildId,
+  int& left,
+  int& top,
+  int& right,
+  int& bottom
+) {
+  if (!Ambition::config.enhancementsAvailable()) {
+    return;
+  }
+
+  const auto offset = BUILDING_DRAWING_OFFSETS.find(_7kaaFirmBuildId);
+  if (offset != BUILDING_DRAWING_OFFSETS.end()) {
+    left += offset->second.left;
+    top += offset->second.top;
+    right += offset->second.left;
+    bottom += offset->second.top;
+  }
 }
 
 int calculateHitbarBaseColour(